+++ indexer/services/comlink/__tests__/controllers/api/v4/compliance-v2-controller.test.ts
@@ -13,10 +13,9 @@ import { RequestMethod } from '../../../../src/types';
 import { logger, stats } from '@dydxprotocol-indexer/base';
 import { redis } from '@dydxprotocol-indexer/redis';
 import { ratelimitRedis } from '../../../../src/caches/rate-limiters';
-import { ComplianceControllerHelper } from '../../../../src/controllers/api/v4/compliance-controller';
 import config from '../../../../src/config';
 import { DateTime } from 'luxon';
-import { ComplianceAction } from '../../../../src/controllers/api/v4/compliance-v2-controller';
+import { ComplianceAction, ComplianceControllerHelper } from '../../../../src/controllers/api/v4/compliance-v2-controller';
 import { ExtendedSecp256k1Signature, Secp256k1 } from '@cosmjs/crypto';
 import { getGeoComplianceReason } from '../../../../src/helpers/compliance/compliance-utils';
 import { isRestrictedCountryHeaders } from '@dydxprotocol-indexer/compliance';
