+++ indexer/services/comlink/src/controllers/api/v4/compliance-v2-controller.ts
@@ -3,18 +3,20 @@ import {
 } from '@cosmjs/crypto';
 import { toBech32 } from '@cosmjs/encoding';
 import { logger, stats, TooManyRequestsError } from '@dydxprotocol-indexer/base';
-import { CountryHeaders, isRestrictedCountryHeaders } from '@dydxprotocol-indexer/compliance';
+import { CountryHeaders, isRestrictedCountryHeaders, ComplianceClientResponse, INDEXER_COMPLIANCE_BLOCKED_PAYLOAD  } from '@dydxprotocol-indexer/compliance';
 import {
   ComplianceReason,
   ComplianceStatus,
   ComplianceStatusFromDatabase,
   ComplianceStatusTable,
+  ComplianceTable,
+  ComplianceDataFromDatabase,
 } from '@dydxprotocol-indexer/postgres';
 import express from 'express';
 import { matchedData } from 'express-validator';
 import { DateTime } from 'luxon';
 import {
-  Controller, Get, Path, Route,
+  Controller, Get, Path, Route, Query,
 } from 'tsoa';
 
 import { getReqRateLimiter } from '../../../caches/rate-limiters';
@@ -29,10 +31,17 @@ import { CheckAddressSchema } from '../../../lib/validation/schemas';
 import { handleValidationErrors } from '../../../request-helpers/error-handler';
 import ExportResponseCodeStats from '../../../request-helpers/export-response-code-stats';
 import { ComplianceRequest, ComplianceV2Response, SetComplianceStatusRequest } from '../../../types';
-import { ComplianceControllerHelper } from './compliance-controller';
+
+import {
+  screenProviderLimiter,
+  screenProviderGlobalLimiter,
+} from '../../../caches/rate-limiters';
+import {ComplianceResponse } from '../../../types';
 
 const router: express.Router = express.Router();
 const controllerName: string = 'compliance-v2-controller';
+const UNCACHED_QUERY_POINTS: number = 1;
+const GLOBAL_RATE_LIMIT_KEY: string = 'screenQueryProviderGlobal';
 
 export enum ComplianceAction {
   ONBOARD = 'ONBOARD',
@@ -45,6 +54,96 @@ const COMPLIANCE_PROGRESSION: Partial<Record<ComplianceStatus, ComplianceStatus>
   [ComplianceStatus.COMPLIANT]: ComplianceStatus.FIRST_STRIKE_CLOSE_ONLY,
   [ComplianceStatus.FIRST_STRIKE]: ComplianceStatus.CLOSE_ONLY,
 };
+
+
+export class ComplianceControllerHelper extends Controller {
+  private ipAddress: string;
+
+  constructor(ipAddress: string) {
+    super();
+    this.ipAddress = ipAddress;
+  }
+
+  @Get('/')
+  async screen(
+    @Query() address: string,
+  ): Promise<ComplianceResponse> {
+    const ageThreshold: DateTime = DateTime.utc().minus({
+      seconds: config.MAX_AGE_SCREENED_ADDRESS_COMPLIANCE_DATA_SECONDS,
+    });
+
+    let complianceData:
+    ComplianceDataFromDatabase | undefined = await ComplianceTable.findByAddressAndProvider(
+      address,
+      complianceProvider.provider,
+    );
+
+    if (complianceData !== undefined) {
+      stats.increment(
+        `${config.SERVICE_NAME}.${controllerName}.compliance_data_cache_hit`,
+        { provider: complianceProvider.provider },
+      );
+    }
+
+    // Immediately return for blocked addresses, do not refresh
+    if (complianceData?.blocked) {
+      return {
+        restricted: true,
+        reason: INDEXER_COMPLIANCE_BLOCKED_PAYLOAD,
+      };
+    }
+
+    if (complianceData === undefined || DateTime.fromISO(complianceData.updatedAt) < ageThreshold) {
+      await checkRateLimit(this.ipAddress);
+
+      if (complianceData === undefined) {
+        stats.increment(
+          `${config.SERVICE_NAME}.${controllerName}.compliance_data_cache_miss`,
+          { provider: complianceProvider.provider },
+        );
+      } else {
+        stats.increment(
+          `${config.SERVICE_NAME}.${controllerName}.refresh_compliance_data_cache`,
+          { provider: complianceProvider.provider },
+        );
+      }
+
+      const response:
+      ComplianceClientResponse = await complianceProvider.client.getComplianceResponse(
+        address,
+      );
+      complianceData = await ComplianceTable.upsert({
+        ...response,
+        provider: complianceProvider.provider,
+        updatedAt: DateTime.utc().toISO(),
+      });
+    }
+
+    return {
+      restricted: complianceData.blocked,
+      reason: complianceData.blocked ? INDEXER_COMPLIANCE_BLOCKED_PAYLOAD : undefined,
+    };
+  }
+}
+
+async function checkRateLimit(ipAddress: string) {
+  try {
+    await Promise.all([
+      screenProviderLimiter.consume(ipAddress, UNCACHED_QUERY_POINTS),
+      screenProviderGlobalLimiter.consume(GLOBAL_RATE_LIMIT_KEY, UNCACHED_QUERY_POINTS),
+    ]);
+  } catch (reject) {
+    if (reject instanceof Error) {
+      logger.error({
+        at: 'rate-limit',
+        message: 'redis error when checking rate limit',
+        reject,
+      });
+    } else {
+      throw new TooManyRequestsError('Too many requests');
+    }
+  }
+}
 
 @Route('compliance')
 class ComplianceV2Controller extends Controller {
